requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.1
rembg>=2.0.55
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
pydantic>=2.5.0
onnxruntime>=1.15.0
